requirements.txt
crewai>=0.108.0
langchain>=0.1.11
langchain-openai>=0.1.2
mcpadapt>=0.1.3
mcp>=1.6.0
openai>=1.13.3
pandas>=2.2.1
python-dotenv>=1.0.1
requests>=2.32.3
boto3>=1.37.24
langchain-core>=0.1.0
pydantic~=2.10.6
jmespath~=1.0.1
pytest~=8.3.5
uv~=0.6.3
ipython~=8.32.0
numpy~=2.2.3
pillow~=11.1.0
lxml~=5.3.1
packaging~=24.2
typing_extensions~=4.12.2
h11~=0.14.0
cryptography~=44.0.1
pip~=25.0.1
attrs~=25.1.0
distro~=1.9.0
protobuf~=5.29.3
filelock~=3.17.0
rsa~=4.9
pyasn1~=0.6.1
cffi~=1.17.1
grpcio~=1.70.0
parso~=0.8.4
jedi~=0.19.2
Pygments~=2.19.1
MarkupSafe~=3.0.2
litellm~=1.60.2
httpx~=0.27.2
pytz~=2025.1
posthog~=3.15.1
chromadb~=0.6.3
rich~=13.9.4
markdown-it-py~=3.0.0
PyYAML~=6.0.2
idna~=3.10
multidict~=6.1.0
propcache~=0.3.0
zipp~=3.21.0
sniffio~=1.3.1
auth0-python~=4.8.1
PyJWT~=2.10.1
aiohttp~=3.11.13
yarl~=1.18.3
urllib3~=2.3.0
isodate~=0.7.2
botocore~=1.37.24
s3transfer~=0.11.4
importlib_metadata~=8.5.0
pyproject_hooks~=1.2.0
build~=1.2.2.post1
click~=8.1.8
httpcore~=1.0.7
zstandard~=0.23.0
certifi~=2025.1.31
mock~=5.1.0
jiter~=0.8.2
json5~=0.10.0
pyarrow~=19.0.1
tqdm~=4.67.1
mdurl~=0.1.2
overrides~=7.7.0
charset-normalizer~=3.4.1
networkx~=3.4.2
jsonpickle~=4.0.2
Jinja2~=3.1.5
regex~=2024.11.6
decorator~=5.2.1
sympy~=1.13.3
mpmath~=1.3.0
typer~=0.15.1
shellingham~=1.5.4
pydantic_core~=2.27.2
tokenizers~=0.21.0
tomli~=2.2.1
tomli_w~=1.2.0
pyvis~=0.3.2
json_repair~=0.39.1
blinker~=1.9.0
appdirs~=1.4.4
fsspec~=2025.2.0
pyasn1_modules~=0.4.1
cachetools~=5.5.2
python-dateutil~=2.9.0.post0
fastapi~=0.115.8
uvicorn~=0.34.0
starlette~=0.45.3
websockets~=15.0
anyio~=4.8.0
orjson~=3.10.15
openpyxl~=3.1.5
SQLAlchemy~=2.0.38
PyPika~=0.48.9
uvloop~=0.21.0
websocket-client~=1.8.0
traitlets~=5.14.3
stack-data~=0.6.3
prompt_toolkit~=3.0.50
wcwidth~=0.2.13
pluggy~=1.5.0
Deprecated~=1.2.18
iniconfig~=2.0.0
pexpect~=4.9.0
aiosignal~=1.3.2
frozenlist~=1.5.0
aiohappyeyeballs~=2.4.6
importlib_resources~=6.5.2
asgiref~=3.8.1
backoff~=2.2.1
arrow~=1.3.0
instructor~=1.7.2
huggingface-hub~=0.29.1
tiktoken~=0.9.0
ptyprocess~=0.7.0
six~=1.17.0
freezegun~=1.5.1
monotonic~=1.6
tzdata~=2025.1
httptools~=0.6.4
watchfiles~=1.0.4
tenacity~=9.0.0
bcrypt~=4.2.1
onnxruntime~=1.20.1
kubernetes~=32.0.1
mmh3~=5.1.0
chroma-hnswlib~=0.7.6
greenlet~=3.1.1
oauthlib~=3.2.2
et_xmlfile~=2.0.0
annotated-types~=0.7.0
chardet~=5.2.0
tornado~=6.4.2
asttokens~=3.0.0
executing~=2.2.0
langsmith~=0.3.11
langchain-text-splitters~=0.3.6
streamlit~=1.42.2
requests-toolbelt~=1.0.0
pure_eval~=0.2.3
pycparser~=2.22
pypdfium2~=4.30.1
wrapt~=1.17.2
docstring_parser~=0.16
jsonref~=1.1.0
referencing~=0.36.2
jsonschema~=4.23.0
jsonpointer~=3.0.0
jsonschema-specifications~=2024.10.1
durationpy~=0.9
requests-oauthlib~=2.0.0
platformdirs~=4.3.6
humanfriendly~=10.0
coloredlogs~=15.0.1
flatbuffers~=25.2.10
jsonpatch~=1.33
pdfplumber~=0.11.5
